DockerBot: read messages until the socket queue is empty
This prevents increasing de-sync between the game and the bot state if
bots have occasional timeouts. In such an event, the bot does not
respond in time and is ignored in that frame. However, it still
processes its data and generates a response, which is stuck in the
socket queue and is read in the next frame. The bot however puts a new
message into the socket queue. This resulted in a permanent processing
delay before this change.