Make x and y of IpcFoodInfo and IpcSegmentInfo useful
Before this change, x and y were given in "world orientation",
meaning that they were translated to a coordinate system where
the origin is the bot's head, but the x and y axes are parallel to
the world coordinate's axes.

This basically rendered x and y useless, because nothing else is given
in that coordinate system, and the bot has no idea about its own rotation.

Modify the values of x and y (i.e. rotate them) to match the coordinate
system of all other values, i.e. where the y axis is the bot's direction.