rust framework: mute dead_code warning on mmap variable
The mmap handle must be kept around for the lifetime of the Api
instance, but it is never read, because all accesses go directly to the
shared memory. Therefore rustc produces a “field is never read” warning,
which will disturb the player, as he can do nothing about it.

Therefore, the warning was muted at this location.